Added ability to invoke commands in public, private (Direct Messages) or both discord channel types

+++ src/com/lucianms/event/ChatHandler.java
@@ -15,16 +15,21 @@ import java.util.regex.Pattern;
  */
 public class ChatHandler {
 
+    public void onDirectMessage(MessageReceivedEvent event) {
+    }
+
     @EventSubscriber
     public void onGuildMessageReceived(MessageReceivedEvent event) {
         String content = event.getMessage().getContent();
-        if (event.getAuthor().getRolesForGuild(event.getGuild()).stream().noneMatch(r -> r.getPermissions().contains(Permissions.MANAGE_MESSAGES))) {
-            for (String words : Discord.getBlacklistedWords()) {
-                Pattern pattern = Pattern.compile(words, Pattern.CASE_INSENSITIVE | Pattern.MULTILINE);
-                Matcher matcher = pattern.matcher(content);
-                if (matcher.find()) {
-                    event.getMessage().delete();
-                    return;
+        if (!event.getChannel().isPrivate()) {
+            if (event.getAuthor().getRolesForGuild(event.getGuild()).stream().noneMatch(r -> r.getPermissions().contains(Permissions.MANAGE_MESSAGES))) {
+                for (String words : Discord.getBlacklistedWords()) {
+                    Pattern pattern = Pattern.compile(words, Pattern.CASE_INSENSITIVE | Pattern.MULTILINE);
+                    Matcher matcher = pattern.matcher(content);
+                    if (matcher.find()) {
+                        event.getMessage().delete();
+                        return;
+                    }
                 }
             }
         }
